refactor: Move 'components' folder into 'views'.
