+++ src/views/components/ImageIndicator.tsx
