+++ src/views/components/ProductSmallPhoto.tsx
